# workspace/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NetOps Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /provisioning/templates/{id}:
    get:
      operationId: getProvisioningTemplate
      tags: [provisioning]
      summary: Get provisioning template by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningTemplate"
        "404":
          description: Not found
...
  /provisioning/preview/export:
    post:
      operationId: exportProvisioningPreview
      tags: [provisioning]
      summary: Export provisioning preview plan (templateId workflow)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisioningPreviewExportRequest"
      responses:
        "200":
          description: Exported preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningPreviewExportResult"
...
components:
  schemas:
...
    ProvisioningParameterField:
      type: object
      required: [type, description]
      properties:
        type:
          type: string
        description:
          type: string
        required:
          type: boolean
        sensitive:
          type: boolean
    ProvisioningTemplate:
      type: object
      required: [id, name, description, vendor, platform, serviceType, parameterSchema, risks, precheckHints, postcheckHints, supported]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        vendor:
          type: string
        platform:
          type: string
        serviceType:
          type: string
        parameterSchema:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ProvisioningParameterField"
        risks:
          type: array
          items:
            type: string
        precheckHints:
          type: array
          items:
            type: string
        postcheckHints:
          type: array
          items:
            type: string
        supported:
          type: boolean
    ProvisioningValidationItem:
      type: object
      required: [name, passed, message]
      properties:
        name:
          type: string
        passed:
          type: boolean
        message:
          type: string
        severity:
          type: string
          enum: [info, warn, error]
    ProvisioningRisk:
      type: object
      required: [code, message, severity]
      properties:
        code:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [info, warn, error]
    ProvisioningPreviewRequest:
      type: object
      required: [deviceId, templateId]
      properties:
        deviceId:
          type: integer
        templateId:
          type: string
        parameters:
          type: object
          additionalProperties: true
        mode:
          type: string
          default: dry_run
        maintenanceWindowStart:
          type: ["string", "null"]
        maintenanceWindowEnd:
          type: ["string", "null"]
        rollbackPlan:
          type: ["string", "null"]
    ProvisioningPreviewResponse:
      type: object
      required:
        - status
        - deviceId
        - templateId
        - serviceType
        - configPreview
        - rollbackPreview
        - executionPlan
        - validations
        - risks
        - precheckHints
        - postcheckHints
        - missingData
        - blockedReasons
        - applyBlocked
      properties:
        status:
          type: string
          enum: [valid, warning, blocked]
        deviceId:
          type: integer
        templateId:
          type: string
        serviceType:
          type: string
        configPreview:
          type: string
        rollbackPreview:
          type: string
        executionPlan:
          type: array
          items:
            type: string
        validations:
          type: array
          items:
            $ref: "#/components/schemas/ProvisioningValidationItem"
        risks:
          type: array
          items:
            $ref: "#/components/schemas/ProvisioningRisk"
        precheckHints:
          type: array
          items:
            type: string
        postcheckHints:
          type: array
          items:
            type: string
        missingData:
          type: array
          items:
            type: string
        blockedReasons:
          type: array
          items:
            type: string
        applyBlocked:
          type: boolean
        applyBlockedReason:
          type: ["string", "null"]
        maintenanceWindow:
          type: ["object", "null"]
          properties:
            start:
              type: ["string", "null"]
            end:
              type: ["string", "null"]
        rollbackPlan:
          type: ["string", "null"]
    ProvisioningPreviewExportRequest:
      allOf:
        - $ref: "#/components/schemas/ProvisioningPreviewRequest"
        - type: object
          required: [format]
          properties:
            format:
              type: string
              enum: [markdown, json]
    ProvisioningPreviewExportResult:
      type: object
      required: [format, content, preview]
      properties:
        format:
          type: string
          enum: [markdown, json]
        content:
          type: string
        preview:
          $ref: "#/components/schemas/ProvisioningPreviewResponse"
...